Move exported types into types.ts and use extends for repeated APIs
