+++ src/client/jupyter/jupyterIntegration.ts
@@ -6,9 +6,8 @@
 
 import { inject, injectable, named } from 'inversify';
 import { dirname } from 'path';
-import { CancellationToken, Disposable, Event, Extension, Memento, Uri } from 'vscode';
-import * as lsp from 'vscode-languageserver-protocol';
-import { ILanguageServerCache, ILanguageServerConnection } from '../activation/types';
+import { CancellationToken, Extension, Memento, Uri } from 'vscode';
+import { ILanguageServerCache } from '../activation/types';
 import { JUPYTER_EXTENSION_ID } from '../common/constants';
 import { InterpreterUri, ModuleInstallFlags } from '../common/installer/types';
 import {
@@ -26,125 +25,17 @@ import { isResource } from '../common/utils/misc';
 import { getDebugpyPackagePath } from '../debugger/extension/adapter/remoteLaunchers';
 import { IEnvironmentActivationService } from '../interpreter/activation/types';
 import { IInterpreterQuickPickItem, IInterpreterSelector } from '../interpreter/configuration/types';
-import {
-    IComponentAdapter,
-    IInterpreterDisplay,
-    IInterpreterService,
-    IInterpreterStatusbarVisibilityFilter,
-} from '../interpreter/contracts';
+import { IComponentAdapter, IInterpreterDisplay, IInterpreterService } from '../interpreter/contracts';
 import { PythonEnvironment } from '../pythonEnvironments/info';
-import { IDataViewerDataProvider, IJupyterUriProvider } from './types';
+import {
+    IDataViewerDataProvider,
+    IJupyterUriProvider,
+    JupyterExtensionApi,
+    JupyterProductToInstall,
+    ProductMapping,
+} from './types';
 import { inDiscoveryExperiment } from '../common/experiments/helpers';
 import { isWindowsStoreInterpreter } from '../pythonEnvironments/discovery/locators/services/windowsStoreInterpreter';
-
-export interface ILanguageServer extends Disposable {
-    readonly connection: ILanguageServerConnection;
-    readonly capabilities: lsp.ServerCapabilities;
-}
-
-/**
- * This allows Python extension to update Product enum without breaking Jupyter.
- * I.e. we have a strict contract, else using numbers (in enums) is bound to break across products.
- */
-export enum JupyterProductToInstall {
-    jupyter = 'jupyter',
-    ipykernel = 'ipykernel',
-    notebook = 'notebook',
-    kernelspec = 'kernelspec',
-    nbconvert = 'nbconvert',
-    pandas = 'pandas',
-}
-
-export const ProductMapping: { [key in JupyterProductToInstall]: Product } = {
-    [JupyterProductToInstall.ipykernel]: Product.ipykernel,
-    [JupyterProductToInstall.jupyter]: Product.jupyter,
-    [JupyterProductToInstall.kernelspec]: Product.kernelspec,
-    [JupyterProductToInstall.nbconvert]: Product.nbconvert,
-    [JupyterProductToInstall.notebook]: Product.notebook,
-    [JupyterProductToInstall.pandas]: Product.pandas,
-};
-
-export type PythonApiForJupyterExtension = {
-    /**
-     * IInterpreterService
-     */
-    onDidChangeInterpreter: Event<void>;
-    /**
-     * IInterpreterService
-     */
-    getInterpreters(resource?: Uri): Promise<PythonEnvironment[]>;
-    /**
-     * IInterpreterService
-     */
-    getActiveInterpreter(resource?: Uri): Promise<PythonEnvironment | undefined>;
-    /**
-     * IInterpreterService
-     */
-    getInterpreterDetails(pythonPath: string, resource?: Uri): Promise<undefined | PythonEnvironment>;
-
-    /**
-     * IEnvironmentActivationService
-     */
-    getActivatedEnvironmentVariables(
-        resource: Resource,
-        interpreter?: PythonEnvironment,
-        allowExceptions?: boolean,
-    ): Promise<NodeJS.ProcessEnv | undefined>;
-    isWindowsStoreInterpreter(pythonPath: string): Promise<boolean>;
-    getSuggestions(resource: Resource): Promise<IInterpreterQuickPickItem[]>;
-    /**
-     * IInstaller
-     */
-    install(
-        product: JupyterProductToInstall,
-        resource?: InterpreterUri,
-        cancel?: CancellationToken,
-    ): Promise<InstallerResponse>;
-    /**
-     * IInstaller
-     */
-    isProductVersionCompatible(
-        product: Product,
-        semVerRequirement: string,
-        resource?: InterpreterUri,
-    ): Promise<ProductInstallStatus>;
-    /**
-     * Returns path to where `debugpy` is. In python extension this is `/pythonFiles/lib/python`.
-     */
-    getDebuggerPath(): Promise<string>;
-    /**
-     * Retrieve interpreter path selected for Jupyter server from Python memento storage
-     */
-    getInterpreterPathSelectedForJupyterServer(): string | undefined;
-    /**
-     * Returns a ILanguageServer that can be used for communicating with a language server process.
-     * @param resource file that determines which connection to return
-     */
-    getLanguageServer(resource?: InterpreterUri): Promise<ILanguageServer | undefined>;
-    /**
-     * Registers a visibility filter for the interpreter status bar.
-     */
-    registerInterpreterStatusFilter(filter: IInterpreterStatusbarVisibilityFilter): void;
-};
-
-export type JupyterExtensionApi = {
-    /**
-     * Registers python extension specific parts with the jupyter extension
-     * @param interpreterService
-     */
-    registerPythonApi(interpreterService: PythonApiForJupyterExtension): void;
-    /**
-     * Launches Data Viewer component.
-     * @param {IDataViewerDataProvider} dataProvider Instance that will be used by the Data Viewer component to fetch data.
-     * @param {string} title Data Viewer title
-     */
-    showDataViewer(dataProvider: IDataViewerDataProvider, title: string): Promise<void>;
-    /**
-     * Registers a remote server provider component that's used to pick remote jupyter server URIs
-     * @param serverProvider object called back when picking jupyter server URI
-     */
-    registerRemoteServerProvider(serverProvider: IJupyterUriProvider): void;
-};
 
 @injectable()
 export class JupyterExtensionIntegration {
